fix: testmm and refactor testsync

+++ Userland/SampleCodeModule/include/lib.h
@@ -37,7 +37,7 @@ int intLength(int num, int minLen);
  * 0's up to minLen
  * */
 
-void intToString(int num, char *buf, int minLen);
+void intToString(int64_t num, char *buf, int minLen);
 
 /**
  * Used to create a new process
